+++ src/main/java/uk/co/caprica/vlcj/player/base/MediaPlayerEventAdapter.java
@@ -85,7 +85,7 @@ public class MediaPlayerEventAdapter implements MediaPlayerEventListener {
     }
 
     @Override
-    public void positionChanged(MediaPlayer mediaPlayer, float newPosition) {
+    public void positionChanged(MediaPlayer mediaPlayer, double newPosition) {
     }
 
     @Override
